Changing and adding
Some things are now translated in english (except for quotes, because I first need to find the original translation for the quotes that have one.

I added a About-Page which will be changed to a separate class and layout, but its actually 3:17 am here and i am tired.

I made the Text of the RecyclerViews to Strings so I can translate into english

I added a Subtitle for the expanded ToolbarLayout which includes the name of the version.

I added special thanks to Yanndroid.