Especially adding
I added the logo of GitHub and Telegram so if i release the app, people will have the possibility to reach me, in case they found an issue or the have quotes for me.

I added a new activity for "About this app"

I re-added the special thanks to Yanndroid
I added special thanks to BlackMesa

I added translation for a few non-quote-Strings
I added Strings to strings.xml for BlackMesa, About this app, etc.

I modified some of the strings